refactor(src): restructured the src folder
