+++ src/core/db/database.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
-import { User } from '@app/components/users/entities/user.entity';
+import { User } from '@app/feature/users/entities/user.entity';
 
 @Module({
   imports: [
