+++ src/core/index.ts
@@ -1,2 +1,2 @@
-export * from './guards';
-export * from './http';
+export * from './middleware';
+export * from '../shared/http';
