+++ src/feature/users/dto/create.user.dto.ts
@@ -1,6 +1,6 @@
 import { ApiProperty } from '@nestjs/swagger';
 
-import { apiResponse } from '@app/components/users/constants/api.response.dto';
+import { apiResponse } from '@app/feature/users/constants/api.response.dto';
 
 export class CreateUserDto {
   @ApiProperty(apiResponse.apiCreateUserFirstNameProperty)
