+++ src/feature/users/repository/user.repository.ts
@@ -1,8 +1,8 @@
 import { UpdateResult } from 'typeorm';
 
-import { CreateUserDto } from '@app/components/users/dto/create.user.dto';
-import { UpdateUserDto } from '@app/components/users/dto/update.user.dto';
-import { User } from '@app/components/users/entities/user.entity';
+import { CreateUserDto } from '@app/feature/users/dto/create.user.dto';
+import { UpdateUserDto } from '@app/feature/users/dto/update.user.dto';
+import { User } from '@app/feature/users/entities/user.entity';
 
 export interface UserRepository {
   findUser(id: string): Promise<User>;
