+++ src/feature/users/users.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
-import { UserDbRepository } from '@app/components/users/repository/db/user.repository';
-import { UsersService } from '@app/components/users/services/users.service';
-import { UsersController } from '@app/components/users/users.controller';
+import { UserDbRepository } from '@app/feature/users/repository/db/user.repository';
+import { UsersService } from '@app/feature/users/services/users.service';
+import { UsersController } from '@app/feature/users/users.controller';
 
 @Module({
   imports: [TypeOrmModule.forFeature([UserDbRepository])],
