+++ src/shared/mailer/email-templates/neosoft_logo.png
