+++ src/shared/mailer/email-templates/nest_logo.png
