+++ src/shared/mailer/mailer-interfaces/interfaces.ts
