+++ test/unit/components/user.repository.spec.ts
@@ -4,8 +4,8 @@ import { AppController_test } from '@test/e2e/user.e2e';
 import { userStub } from '@test/mock/user.stub';
 
 import { AppModule } from '@app/app.module';
-import { User } from '@app/components/users/entities/user.entity';
-import { UserDbRepository } from '@app/components/users/repository/db/user.repository';
+import { User } from '@app/feature/users/entities/user.entity';
+import { UserDbRepository } from '@app/feature/users/repository/db/user.repository';
 
 describe('Testing UserDbRepository', () => {
   let app: INestApplication;
