+++ test/unit/components/users.service.spec.ts
@@ -2,11 +2,11 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { getRepositoryToken } from '@nestjs/typeorm';
 import { userStub } from '@test/mock/user.stub';
 
-import { CreateUserDto } from '@app/components/users/dto/create.user.dto';
-import { UpdateUserDto } from '@app/components/users/dto/update.user.dto';
-import { User } from '@app/components/users/entities/user.entity';
-import { UserDbRepository } from '@app/components/users/repository/db/user.repository';
-import { UsersService } from '@app/components/users/services/users.service';
+import { CreateUserDto } from '@app/feature/users/dto/create.user.dto';
+import { UpdateUserDto } from '@app/feature/users/dto/update.user.dto';
+import { User } from '@app/feature/users/entities/user.entity';
+import { UserDbRepository } from '@app/feature/users/repository/db/user.repository';
+import { UsersService } from '@app/feature/users/services/users.service';
 
 jest.mock('bcrypt', () => ({
   genSalt: jest.fn().mockResolvedValue('Salt'),
