+++ test/unit/core/typeorm.spec.ts
@@ -3,8 +3,8 @@ import { ConfigService } from '@nestjs/config';
 import { Test, TestingModule } from '@nestjs/testing';
 import { Connection, createConnection } from 'typeorm';
 
-import { User } from '@app/components/users/entities/user.entity';
 import { CoreModule } from '@app/core/module';
+import { User } from '@app/feature/users/entities/user.entity';
 
 describe('Testing Typeorm connection', () => {
   let app: INestApplication;
